Broke out Ballot, Contest, and Printable into separate files
